Copy Wallaroo Pony TCP classes to BYOI
This sets up the initial work for Build Your Own Integration. There
are now classes copied from the TCP classes that can be evolved to
support the features that are needed for custom source and sink
integration. This work will eventually make it easier for us and our
customers to build new integrations quickly.