Remove specialized RouterBuilders
There is no longer any need for a special TypedRoute.
Replacing it with an untyped DirectRoute allows a
lot more flexibility when it comes to updating routes.
We can now create routes w/o any further information
by checking whether the Consumer is a boundary or
not. If it is, we create a BoundaryRoute. Otherwise
we create a DirectRoute.