Adopt the connector naming for the API
This is an attempt to align with the general terminology used withing
the stream processing community. BYOI should no longer appear in
the code.

I've made the demonstration scripts available as reusable connectors.
They're focussed on being self-contained outside of third-party
library support for the integration and anything in the wallaroo
python library (specifically the code in experimental).